# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/product/create": {
            "post": {
                "description": "创建一个新的商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "创建商品",
                "parameters": [
                    {
                        "description": "商品信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.productCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商品id",
                        "schema": {
                            "$ref": "#/definitions/handler.idReq"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/product/delete": {
            "delete": {
                "description": "删除商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "删除商品",
                "parameters": [
                    {
                        "description": "商品id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.idReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/product/get": {
            "get": {
                "description": "获取单个商品信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "获取商品信息",
                "parameters": [
                    {
                        "description": "商品id",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.idReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商品信息",
                        "schema": {
                            "$ref": "#/definitions/handler.simpleProduct"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/product/list": {
            "get": {
                "description": "获取指定类别的商品列表，不指定类别返回全部商品的列表",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "商品列表",
                "parameters": [
                    {
                        "description": "商品列表信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.productListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商品列表",
                        "schema": {
                            "$ref": "#/definitions/handler.productListReq"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/product/search": {
            "get": {
                "description": "搜索在名字和描述中含有关键词的商品",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "商品搜索",
                "parameters": [
                    {
                        "description": "搜索关键词",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.searchReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "商品信息",
                        "schema": {
                            "$ref": "#/definitions/handler.simpleProduct"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/product/update": {
            "put": {
                "description": "更新商品信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "product"
                ],
                "summary": "更新商品信息",
                "parameters": [
                    {
                        "description": "商品信息",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.simpleProduct"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/test/auth/delete": {
            "get": {
                "description": "删除一个token，即退出该用户在该机器上的登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test/auth"
                ],
                "summary": "删除token",
                "parameters": [
                    {
                        "description": "token",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_handler.tokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    }
                }
            }
        },
        "/test/auth/deleteall": {
            "get": {
                "description": "删除一个用户的所有token，即退出该用户在所有机器上的登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test/auth"
                ],
                "summary": "删除所有token",
                "parameters": [
                    {
                        "description": "token",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_handler.useridInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    }
                }
            }
        },
...
        "/test/auth/verify": {
            "get": {
                "description": "验证一个token是否有效，以及属于哪个用户。会自动为token续期",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test/auth"
                ],
                "summary": "验证token",
                "parameters": [
                    {
                        "description": "token",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/test_handler.tokenInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "结果",
                        "schema": {
                            "$ref": "#/definitions/test_handler.verifyOutput"
                        }
                    },
                    "400": {
                        "description": "请求格式错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/test_handler.errorReturn"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.errorReturn": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.idReq": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "handler.productCreateReq": {
            "type": "object",
            "required": [
                "description",
                "name",
                "picture",
                "price"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "handler.productListReq": {
            "type": "object",
            "required": [
                "page",
                "page_size"
            ],
            "properties": {
                "category_name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                }
            }
        },
...
        "handler.searchReq": {
            "type": "object",
            "required": [
                "query"
            ],
            "properties": {
                "query": {
                    "type": "string"
                }
            }
        },
        "handler.simpleProduct": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "picture": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "test_handler.errorReturn": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "test_handler.tokenInput": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "test_handler.useridInput": {
            "type": "object",
            "required": [
                "userid"
            ],
            "properties": {
                "userid": {
                    "type": "integer"
                }
            }
        },
        "test_handler.verifyOutput": {
            "type": "object",
            "properties": {
                "res": {
                    "type": "boolean"
                },
                "userid": {
                    "type": "integer"
                }
            }
        }
    }
}